move local socket discovery into /local sub-package
