+++ cmd/cli/event.go
@@ -6,6 +6,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/infrakit/pkg/discovery"
+	"github.com/docker/infrakit/pkg/discovery/local"
 	metadata_template "github.com/docker/infrakit/pkg/plugin/metadata/template"
 	"github.com/docker/infrakit/pkg/rpc/client"
 	event_rpc "github.com/docker/infrakit/pkg/rpc/event"
@@ -193,7 +194,7 @@ func eventCommand(plugins func() discovery.Plugins) *cobra.Command {
 
 			log.Infof("Using %v for rendering view.", templateURL)
 			engine, err := template.NewTemplate(templateURL, template.Options{
-				SocketDir: discovery.Dir(),
+				SocketDir: local.Dir(),
 			})
 			if err != nil {
 				return err
