+++ cmd/cli/template.go
@@ -6,6 +6,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/infrakit/pkg/discovery"
+	"github.com/docker/infrakit/pkg/discovery/local"
 	metadata_template "github.com/docker/infrakit/pkg/plugin/metadata/template"
 	"github.com/docker/infrakit/pkg/template"
 	"github.com/spf13/cobra"
@@ -22,7 +23,7 @@ func templateCommand(plugins func() discovery.Plugins) *cobra.Command {
 
 			log.Infof("Using %v for reading template\n", templateURL)
 			engine, err := template.NewTemplate(templateURL, template.Options{
-				SocketDir: discovery.Dir(),
+				SocketDir: local.Dir(),
 			})
 			if err != nil {
 				return err
