+++ cmd/manager/etcd.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/coreos/etcd/clientv3"
 	"github.com/docker/go-connections/tlsconfig"
-	"github.com/docker/infrakit/pkg/discovery"
+	"github.com/docker/infrakit/pkg/discovery/local"
 	etcd_leader "github.com/docker/infrakit/pkg/leader/etcd/v3"
 	etcd_store "github.com/docker/infrakit/pkg/store/etcd/v3"
 	etcd "github.com/docker/infrakit/pkg/util/etcd/v3"
@@ -66,7 +66,7 @@ func etcdEnvironment(getConfig func() config) *cobra.Command {
 			return err
 		}
 
-		plugins, err := discovery.NewPluginDiscovery()
+		plugins, err := local.NewPluginDiscovery()
 		if err != nil {
 			return err
 		}
